requirements.txt
bokeh==2.4.1
numpy
pandas
streamlit
umap-learn
pacmap
hdbscan
requests